Create catcher for JSON errors (#5) (#12)
* Catch empty JSON requests (#5)

Hoping I proposed this correctly (first time). This should now gracefully catch whenever a JSON request returns empty. If wanted, this can now be passed on through each request to give the user the ability to handle when an error occurs. As of now, it returns the default value of the object requested on error. As a side note, I also slightly cleaned up of the code presented in this file.

* Separate ApiRequest into its own file.

Just a file transfer.

* Create catcher for JSON errors (#5)

This class is used to help gracefully catch JSON errors on requests.